// src/sdk/generated/mma.d.ts
export interface components {
    schemas: {
// ...
        EventDetail: {
            /** @description The details of the fights */
            Fights?: components["schemas"]["Fight"][];
            /** @description The unique ID of this event */
            EventId?: number;
            /** @description The unique ID of this event's MMA league */
            LeagueId?: number;
            /** @description The full name of the event */
            Name?: string | null;
            /** @description The short name of the event */
            ShortName?: string | null;
            /** @description The season the event took place */
            Season?: number | null;
            /** @description The date of the event */
            Day?: string | null;
            /** @description The start time and day of the event */
            DateTime?: string | null;
            /** @description Indicates the event's status. Possible values include: Scheduled, In Progress, Final, Suspended, Postponed, Canceled */
            Status?: string | null;
            /** @description Indicates if the event is active */
            Active?: boolean | null;
        };
        Fight: {
            /** @description The unique ID of this fight */
            FightId?: number;
            /** @description The order of this fight on the fight card */
            Order?: number | null;
            /** @description Indicates the event's status. Possible values include: Scheduled; In Progress; Final; Suspended; Postponed; Canceled; NULL */
            Status?: string | null;
            /** @description The weight class this fight is taking place at */
            WeightClass?: string | null;
            /** @description What segment of the card this fight is on (Main Card, Prelims, Early Prelims) */
            CardSegment?: string | null;
            /** @description The number of referee for this fight */
            Referee?: string | null;
            /** @description The number of rounds for this fight */
            Rounds?: number | null;
            /** @description The time on the clock in seconds when the fight ended */
            ResultClock?: number | null;
            /** @description The round when the fight ended */
            ResultRound?: number | null;
            /** @description The way in which this fight ended (Possible values: No Contest; Decision; Submission; KO/TKO; Decision - Majority; Decision - Unanimous; Decision - Split; Draw; DQ) */
            ResultType?: string | null;
            /** @description The unique ID of the fighter who won the fight */
            WinnerId?: number | null;
            /** @description The fighters competing in this fight and their info (FighterID, FirstName, LastName, PreFightWins, PreFightLosses, PreFightDraws, PreFightNoContests, Winner, Moneyline) */
            Fighters?: components["schemas"]["FighterInfo"][];
            /** @description Indicates the fight is still active */
            Active?: boolean | null;
        };
        FighterInfo: {
            /** @description The unique ID of this fighter */
            FighterId?: number | null;
            /** @description The first name of the fighter */
            FirstName?: string | null;
            /** @description The last name of the fighter */
            LastName?: string | null;
            /** @description The fighter's win total prior to the fight */
            PreFightWins?: number | null;
            /** @description The fighter's loss total prior to the fight */
            PreFightLosses?: number | null;
            /** @description The fighter's draw total prior to the fight */
            PreFightDraws?: number | null;
            /** @description The fighter's no contest total prior to the fight */
            PreFightNoContests?: number | null;
            /** @description Indicates if this fighter won the fight */
            Winner?: boolean | null;
            /** @description Moneyline for the fighter's odds to win the fight */
            Moneyline?: number | null;
            /** @description Indicates the fighter is still listed for this fight */
            Active?: boolean | null;
        };
        Event: {
            /** @description The unique ID of this event */
            EventId?: number;
            /** @description The unique ID of this event's MMA league */
            LeagueId?: number;
            /** @description The full name of the event */
            Name?: string | null;
            /** @description The short name of the event */
            ShortName?: string | null;
            /** @description The season the event took place */
            Season?: number | null;
            /** @description The date of the event */
            Day?: string | null;
            /** @description The start time and day of the event */
            DateTime?: string | null;
            /** @description Indicates the event's status. Possible values include: Scheduled, In Progress, Final, Suspended, Postponed, Canceled */
            Status?: string | null;
            /** @description Indicates if the event is active */
            Active?: boolean | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}